+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `_all` and `default_fields` options
 - Added global `index_prefix` option
 - Added `wait` option to async reindex
+- Added `model_includes` option
 - Raise error for `reindex_status` when Redis not configured
 
 ## 2.3.1