+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `wait` option to async reindex
 - Added `model_includes` option
 - Raise error for `reindex_status` when Redis not configured
+- Warn when incompatible options used with `body` option
 
 ## 2.3.1